Fix ExpiredTokenException Errors - GH #220
Move creds and boto3 session construction into AwsLimitChecker to avoid
session timeouts in long-lived sessions.

Fixes GH issue #220. See thread for detailed discussion of this change.

Thanks to @jantman for all the help on this.

----

This contribution is being made under the same license as the
awslimitchecker project (AGPL, or any subsequent version of that
license as adopted by awslimitchecker), and may perpetually be
included in and distributed with awslimitchecker.
